Add QSPI support, phase 1.
This is about supporting the single/extended mode of the QSPI module in stock acces  and add the DMA support for this access mode.
This has been tested and allow up to 3MB/s with only 22ns of inter-byte time delay. This lead to a 88% bus load. This is great.