QSPI Support : deinit QSPI on end.
When the upload is ended, the QSPI unit is disabled (deinit) and the associated IOPorts are restored in RESET state (INPUT, 1).
This is mandatory as this unit shares its I/O with the SPI unit on the board.
BTW, this should be done for all the I/O units used by the bootstrap.